Store the maximum number of code hosts with license data.
Akin to the maximum user count supported, this configures how many external services can be added for the current license. It should behave in a backwards-compatible way, where 0 = an unlimited number of code hosts (which is what we support today I believe).

As noted in the comments, this also increments the formatVersion since a new field has been introduced.